% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_mc.R
\name{benchmark_mc}
\alias{benchmark_mc}
\title{Run multicore benchmarks}
\usage{
benchmark_mc(runs = 3, verbose = TRUE, cores = NULL)
}
\arguments{
\item{runs}{Number of times to run the test. Default 3.}

\item{verbose}{Default TRUE.}

\item{cores}{Number of cores to try to use. Default is \code{detectCores}}
}
\description{
This function runs a set of multicore benchmarks, which should be suitable for most
machines. It runs a collection of matrix benchmark functions
\itemize{
\item \code{benchmark_prog}
\item \code{benchmark_matrix_cal}
\item \code{benchmark_matrix_fun}
}
To view the list of benchmarks, see \code{get_available_benchmarks}.
}
\examples{
## Benchmark your system
\dontrun{
res = benchmark_std(3)

## Plot results
plot(res)
}
}
