% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_matrix_functions.R
\name{bm_matrix_fft}
\alias{bm_matrix_cholesky}
\alias{bm_matrix_determinant}
\alias{bm_matrix_eigen}
\alias{bm_matrix_fft}
\alias{bm_matrix_inverse}
\title{Matrix function benchmarks}
\usage{
bm_matrix_fft(runs = 3, verbose = FALSE)

bm_matrix_eigen(runs = 3, verbose = FALSE)

bm_matrix_determinant(runs = 3, verbose = FALSE)

bm_matrix_cholesky(runs = 3, verbose = FALSE)

bm_matrix_inverse(runs = 3, verbose = FALSE)
}
\arguments{
\item{runs}{Number of times to run the test. Default 3.}

\item{verbose}{Default TRUE.}
}
\description{
A collection of matrix benchmark functions
\itemize{
\item FFT over 2400000 random values.
\item Eigenvalues of a 640x640 random matrix.
\item Determinant of a 2500x2500 random matrix.
\item Cholesky decomposition of a 3000x3000 matrix.
\item Inverse of a 1600x1600 random matrix.
}
These benchmarks have been developed by many authors. See http://r.research.att.com/benchmarks/R-benchmark-25.R
for a complete history.
}
\references{
http://r.research.att.com/benchmarks/R-benchmark-25.R
}

