% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_matrix_calculations.R
\name{bm_matrix_manip}
\alias{bm_matrix_cross_product}
\alias{bm_matrix_lm}
\alias{bm_matrix_manip}
\alias{bm_matrix_power}
\alias{bm_matrix_sort}
\title{Matrix calculations}
\usage{
bm_matrix_manip(runs = 3, verbose = FALSE)

bm_matrix_power(runs = 3, verbose = FALSE)

bm_matrix_sort(runs = 3, verbose = FALSE)

bm_matrix_cross_product(runs = 3, verbose = FALSE)

bm_matrix_lm(runs = 3, verbose = FALSE)
}
\description{
Matrix calculations
}

