% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_prog_mc.R
\name{bm_prog_fib_mc}
\alias{bm_prog_fib_mc}
\alias{bm_prog_hilbert_mc}
\alias{bm_prog_gcd_mc}
\alias{bm_prog_toeplitz_mc}
\alias{bm_prog_escoufier_mc}
\title{Programming benchmarks}
\usage{
bm_prog_fib_mc(runs = 3, verbose = TRUE, cores = NULL)

bm_prog_hilbert_mc(runs = 3, verbose = TRUE, cores = NULL)

bm_prog_gcd_mc(runs = 3, verbose = TRUE, cores = NULL)

bm_prog_toeplitz_mc(runs = 3, verbose = TRUE, cores = NULL)

bm_prog_escoufier_mc(runs = 3, verbose = TRUE, cores = NULL)
}
\arguments{
\item{runs}{Number of times to run the test. Default 3.}

\item{verbose}{Default TRUE.}

\item{cores}{Number of cores to try to use. Default is \code{detectCores}}
}
\description{
A collection of matrix programming benchmark functions
\itemize{
\item 3,500,000 Fibonacci numbers calculation (vector calc).
\item Creation of a 3500x3500 Hilbert matrix (matrix calc).
\item Grand common divisors of 1,000,000 pairs (recursion).
\item Creation of a 1600x1600 Toeplitz matrix (loops).
\item Escoufier's method on a 60x60 matrix (mixed).
}
These benchmarks have been developed by many authors. See http://r.research.att.com/benchmarks/R-benchmark-25.R
for a complete history. The function \code{benchmark_prog()} runs the five \code{bm} functions.
}
