% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{results}
\alias{results}
\title{Benchmarking results}
\format{A data frame}
\description{
A summary of past benchmarks. The full data set will be 
made available in a companion package
}

