% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{get_available_benchmarks}
\alias{get_available_benchmarks}
\title{Available benchmarks}
\usage{
get_available_benchmarks()
}
\description{
The function returns the available benchmarks
}
\examples{
get_available_benchmarks()
}

