% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpu.R
\name{get_cpu}
\alias{get_cpu}
\title{CPU Description}
\usage{
get_cpu()
}
\description{
Attempt to extracting the CPU model on the current host. This is OS 
specific:
\itemize{
\item Linux: \code{/proc/cpuinfo}
\item Apple: \code{sysctl -n}
\item Windows: \code{wmic cpu}
}
}
\examples{
## Return the machine CPU
get_cpu()
}

