% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{get_ram}
\alias{get_ram}
\alias{print.bytes}
\title{Get the amount of RAM}
\usage{
get_ram()

print.bytes(x, digits = 3, ...)
}
\description{
Extracting the amount of RAM is OS specific and hence messy.
\itemize{
\item Linux: \code{proc/meminfo}
\item Apple: \code{system_profiler -detailLevel mini}
\item Windows: \code{memory.size()}
\item Solaris: \code{prtconf}
}
}
\examples{
get_ram()
}
\references{
The print.bytes function was taken from the pryr package.
}

