% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot.ben_results}
\alias{plot.ben_results}
\title{Compare results to past tests}
\usage{
\method{plot}{ben_results}(x, test_group = NULL,
  byte_optimize = get_byte_compiler(), log = "y", ...)
}
\arguments{
\item{x}{The output from a \code{benchmark_*} call.}

\item{test_group}{Default \code{NULL}. The default behaviour is to average over all tests.}

\item{byte_optimize}{The default behaviour is to compare your results with results that use the same 
byte_optimized setting. To use all results, set to \code{NULL}.}

\item{log}{By default the y axis is plotted on the log scale. To change, set the 
the argument equal to the empty parameter string, \code{""}.}

\item{...}{Arguments to be passed to other downstream methods.}
}
\description{
Plotting
}
\examples{
data(sample_results)
plot(sample_results)
plot(sample_results, byte_optimze=NULL)
}

