% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_past}
\alias{plot_past}
\title{Plot past results}
\usage{
plot_past(byte_optimize = NULL)
}
\arguments{
\item{byte_optimize}{Default \code{NULL}. The default behaviour is to plot all results.
To plot only the byte optimized results, set to \code{TRUE}, otherwise \code{FALSE}}
}
\description{
Plot the previous benchmarks. This function creates two plots. The first showing
time in seconds, the other relative time. The data set used is \code{data(results)}.
}

