% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_results.R
\name{rank_results}
\alias{rank_results}
\title{Benchmark rankings}
\usage{
rank_results(results, test_group = NULL,
  byte_optimize = get_byte_compiler(), verbose = TRUE)
}
\arguments{
\item{results}{Benchmark results. Probably obtained from 
\code{benchmark_std()}.}

\item{test_group}{Default \code{NULL}. The default behaviour is to average over all tests.}

\item{byte_optimize}{The default behaviour is to compare your results with results that use the same 
byte_optimized setting. To use all results, set to \code{NULL}.}

\item{verbose}{Default TRUE.}
}
\description{
Comparison with past results.
}

