% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_results.R
\name{upload_results}
\alias{upload_results}
\title{Upload benchmark results}
\usage{
upload_results(results, url = "http://www.mas.ncl.ac.uk/~ncsg3/form.php",
  sys_info = TRUE, platform_info = TRUE, r_version = TRUE, ram = TRUE,
  cpu = TRUE, byte_compiler = TRUE, linear_algebra = TRUE,
  installed_packages = TRUE)
}
\arguments{
\item{results}{Benchmark results. Probably obtained from \code{benchmark_all()}.}

\item{url}{The location of where to upload the results.}

\item{sys_info}{Default \code{TRUE}.}

\item{platform_info}{Default \code{TRUE}.}

\item{r_version}{Default \code{TRUE}.}

\item{ram}{Default \code{TRUE}.}

\item{cpu}{Default \code{TRUE}.}

\item{byte_compiler}{Default \code{TRUE}.}

\item{linear_algebra}{Default \code{TRUE}.}

\item{installed_packages}{Default \code{TRUE}.}
}
\description{
This function uploads the benchmarking results. These results will then be incorparated
in future versions of the package.
}

