% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_io.R, R/benchmarks.R
\name{bm_io}
\alias{benchmark_io}
\alias{bm_io}
\title{IO benchmarks}
\usage{
bm_io(runs = 3, verbose = TRUE)

benchmark_io(runs = 3, verbose = TRUE)
}
\arguments{
\item{runs}{Number of times to run the test. Default 3.}

\item{verbose}{Default TRUE.}
}
\description{
Benchmarking reading and writing a csv file (containing 2e6 random numbers).
The tests are essentially \code{write.csv(x)} and \code{read.csv(...)} where \code{x} is a data frame.
}

