% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_all.R
\name{benchmark_std}
\alias{benchmark_all}
\alias{benchmark_std}
\alias{get_benchmarks}
\title{Run standard benchmarks}
\usage{
benchmark_std(runs = 3, verbose = TRUE)

benchmark_all(runs = 3, verbose = TRUE)

get_benchmarks()
}
\arguments{
\item{runs}{Number of times to run the test. Default 3.}

\item{verbose}{Default TRUE.}
}
\description{
This functions runs the standard benchmarks, which should be suitable for most
machines
To view the list of benchmarks, see \code{get_benchmarks}.
}

