% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkme-package.R
\docType{package}
\name{benchmarkme-package}
\alias{benchmarkme-package}
\alias{benchmarkme}
\title{The benchmarkme package}
\description{
Benchmark your CPU and compare against other CPUs. Also provides
functions for obtaining system specifications, such as
RAM, CPU type, and R version.
}
\examples{
## Benchmark your system and compare
\dontrun{
res = benchmark_std()
upload_results(res)
plot(res)
}
}
\seealso{
\url{https://github.com/csgillespie/benchmarkme}
}
\author{
\email{csgillespie@gmail.com}
}
\keyword{package}
