% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_matrix_calc_mc.R
\name{bm_matrix_cal_manip_mc}
\alias{bm_matrix_cal_manip_mc}
\alias{bm_matrix_cal_power_mc}
\alias{bm_matrix_cal_sort_mc}
\alias{bm_matrix_cal_cross_product_mc}
\alias{bm_matrix_cal_lm_mc}
\title{Matrix calculation benchmarks}
\usage{
bm_matrix_cal_manip_mc(runs = 3, verbose = TRUE, cores = NULL)

bm_matrix_cal_power_mc(runs = 3, verbose = TRUE, cores = NULL)

bm_matrix_cal_sort_mc(runs = 3, verbose = TRUE, cores = NULL)

bm_matrix_cal_cross_product_mc(runs = 3, verbose = TRUE, cores = NULL)

bm_matrix_cal_lm_mc(runs = 3, verbose = TRUE, cores = NULL)
}
\arguments{
\item{runs}{Number of times to run the test. Default 3.}

\item{verbose}{Default TRUE.}

\item{cores}{Number of cores to try to use. Default is \code{detectCores}}
}
\description{
A collection of matrix benchmark functions aimed at
assessing the calculation speed.
\itemize{
\item Creation, transp., deformation of a 2500x2500 matrix.
\item 2500x2500 normal distributed random matrix ^1000.
\item Sorting of 7,000,000 random values.
\item 2500x2500 cross-product matrix (b = a' * a)
\item Linear regr. over a 3000x3000 matrix.
}
These benchmarks have been developed by many authors. See http://r.research.att.com/benchmarks/R-benchmark-25.R
for a complete history. The function \code{benchmark_matrix_cal()} runs the five \code{bm} functions.
}
\references{
http://r.research.att.com/benchmarks/R-benchmark-25.R
}
