% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_matrix_functions.R
\name{bm_matrix_fft}
\alias{bm_matrix_cholesky}
\alias{bm_matrix_determinant}
\alias{bm_matrix_eigen}
\alias{bm_matrix_fft}
\alias{bm_matrix_inverse}
\title{FFT benchmark}
\usage{
bm_matrix_fft(runs = 3, verbose = FALSE)

bm_matrix_eigen(runs = 3, verbose = FALSE)

bm_matrix_determinant(runs = 3, verbose = FALSE)

bm_matrix_cholesky(runs = 3, verbose = FALSE)

bm_matrix_inverse(runs = 3, verbose = FALSE)
}
\description{
A collection of matrix benchmark functions
}

