% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_matrix_calculations.R
\name{bm_matrix_manip}
\alias{bm_matrix_cross_product}
\alias{bm_matrix_lm}
\alias{bm_matrix_manip}
\alias{bm_matrix_power}
\alias{bm_matrix_sort}
\title{Matrix function benchmarks}
\usage{
bm_matrix_manip(runs = 3, verbose = FALSE)

bm_matrix_power(runs = 3, verbose = FALSE)

bm_matrix_sort(runs = 3, verbose = FALSE)

bm_matrix_cross_product(runs = 3, verbose = FALSE)

bm_matrix_lm(runs = 3, verbose = FALSE)
}
\arguments{
\item{runs}{Number of times to run the test. Default 3.}

\item{verbose}{Default TRUE.}
}
\description{
A collection of matrix benchmark functions aimed at
assessing the calculation speed.
\itemize{
\item Creation, transp., deformation of a 5000x5000 matrix.
\item 3500x3500 normal distributed random matrix ^1000.
\item Sorting of 7,000,000 random values.
\item Cholesky decomposition of a 3000x3000 matrix.
\item Linear regr. over a 3000x3000 matrix.
}
These benchmarks have been developed by many authors. See http://r.research.att.com/benchmarks/R-benchmark-25.R
for a complete history.
}
\references{
http://r.research.att.com/benchmarks/R-benchmark-25.R
}

