% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_programming.R
\name{bm_prog_fib}
\alias{bm_prog_escoufier}
\alias{bm_prog_fib}
\alias{bm_prog_gcd}
\alias{bm_prog_hilbert}
\alias{bm_prog_toeplitz}
\title{Programming benchmarks}
\usage{
bm_prog_fib(runs = 3, verbose = FALSE)

bm_prog_hilbert(runs = 3, verbose = FALSE)

bm_prog_gcd(runs = 3, verbose = FALSE)

bm_prog_toeplitz(runs = 3, verbose = FALSE)

bm_prog_escoufier(runs = 3, verbose = FALSE)
}

