% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_byte_compiler.R
\name{get_byte_compiler}
\alias{get_byte_compiler}
\title{Byte compiler status}
\usage{
get_byte_compiler()
}
\value{
An integer indicating if byte compiling has been turn on. See \code{?compiler} for
details.
}
\description{
Attempts to detect if byte compiling or JIT has been used on the package.
}
\details{
For R 3.5.0 all packages are byte compiled. Before 3.5.0 it was messy. 
Sometimes the user would turn it on via JIT, or ByteCompiling the package. On top of that
R 3.4.X(?) was byte compiled, but R 3.4.Y(?) was, not fully optimised!!! What this means is
don't trust historical results!
}
\examples{
## Detect if you use byte optimization
get_byte_compiler()
}
