% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_byte_compiler.R
\name{get_byte_compiler}
\alias{get_byte_compiler}
\title{Byte compiler status}
\usage{
get_byte_compiler()
}
\value{
An integer indicating if byte compiling has been turn on. See \code{?compiler} for
details.
}
\description{
Attempts to detect if byte compiling has been used on the package. There isn't a fool proof
way of detecting if the the package has been benchmarked.
}

