% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpu.R
\name{get_cpu}
\alias{get_cpu}
\title{CPU Description}
\usage{
get_cpu()
}
\description{
Extracting the amount of RAM is OS specific and hence messy.
\itemize{
\item Linux: \code{/proc/cpuinfo}
\item Apple: \code{sysctl -n}
\item Windows: \code{??}
}
}

