% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{get_datatable}
\alias{get_datatable}
\title{Interactive table of results}
\usage{
get_datatable(results)
}
\arguments{
\item{results}{Benchmark results. Probably obtained from \code{benchmark_all()}.}
}
\description{
Compare your results against past results. Your results is 
shown in Orange.
}

