% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blas_library.R
\name{get_linear_algebra}
\alias{get_linear_algebra}
\title{Get BLAS and LAPACK libraries}
\usage{
get_linear_algebra()
}
\description{
This currently only works under Linux. For other OS, \code{NA}'s 
are returned.
}
\details{
Calls `Sys.getpid()` and greps blas/lapack
}

