% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linear_algebra.R
\name{get_linear_algebra}
\alias{get_linear_algebra}
\title{Get BLAS and LAPACK libraries
Extract the the blas/lapack from \code{sessionInfo()}}
\usage{
get_linear_algebra()
}
\description{
Get BLAS and LAPACK libraries
Extract the the blas/lapack from \code{sessionInfo()}
}
