% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ram.R
\name{get_ram}
\alias{get_ram}
\title{Get the amount of RAM}
\usage{
get_ram()
}
\description{
Extracting the amount of RAM is OS specific and hence messy.
\itemize{
\item Linux: \code{proc/meminfo}
\item Apple: \code{system_profiler -detailLevel mini}
\item Windows: \code{memory.size()}
\item Solaris: \code{prtconf}
}
}
\examples{
## Return (and pretty print) the amount of RAM
get_ram()
}

