% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ram.R
\name{get_ram}
\alias{get_ram}
\title{Get the amount of RAM}
\usage{
get_ram()
}
\description{
Attempt to extract the amount of RAM on the current machine. This is OS 
specific:
\itemize{
\item Linux: \code{proc/meminfo}
\item Apple: \code{system_profiler -detailLevel mini}
\item Windows: \code{memory.size()}
\item Solaris: \code{prtconf}
}
A value of \code{NA} is return if it isn't possible to determine the amount of RAM.
}
\examples{
## Return (and pretty print) the amount of RAM
get_ram()
}
\references{
The \code{print.bytes} function was taken from the \pkg{pryr} package.
}

