% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sys_details.R
\name{get_sys_details}
\alias{get_sys_details}
\title{General system information}
\usage{
get_sys_details()
}
\value{
A list
}
\description{
The \code{get_sys_info} returns general system level information as a list. Each element of the list
is contains the output from: 
\itemize{
\item \code{Sys.info()};
\item \code{get_platform_info()};
\item \code{get_r_version()};
\item \code{get_ram()};
\item \code{get_cpu()};
\item \code{get_byte_compiler()};
\item \code{get_linear_algebra()};
\item \code{installed.packages()}
\item The current date.
}
}

