% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_results.R
\name{rank_results}
\alias{rank_results}
\title{Benchmark rankings}
\usage{
rank_results(results, verbose = TRUE)
}
\arguments{
\item{results}{Benchmark results. Probably obtained from 
\code{benchmark_std()}.}

\item{verbose}{Default TRUE.}
}
\description{
Comparison with past resultss
}

