% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_results.R
\name{rank_results}
\alias{rank_results}
\title{Benchmark rankings}
\usage{
rank_results(results, test_group = unique(results$test_group),
  verbose = TRUE)
}
\arguments{
\item{results}{Benchmark results. Probably obtained from 
\code{benchmark_std()} or \code{benchmark_io()}.}

\item{test_group}{Default \code{unique(x$test_group)}. 
The default behaviour is select the groups from your benchmark results.}

\item{verbose}{Default TRUE.}
}
\description{
Comparison with past results.
}
