% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R, R/plot_results.R
\docType{import}
\name{reexports}
\alias{get_datatable_past}
\alias{plot_past}
\alias{reexports}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{benchmarkmeData}{\code{\link[benchmarkmeData]{get_datatable_past}}, \code{\link[benchmarkmeData]{plot_past}}}
}}
\keyword{internal}

