% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R, R/rank_results.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plot_past}
\alias{is_blas_optimize}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{benchmarkmeData}{\code{\link[benchmarkmeData]{plot_past}}, \code{\link[benchmarkmeData]{is_blas_optimize}}}
}}

