% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_results.R
\name{upload_results}
\alias{upload_results}
\title{Upload benchmark results}
\usage{
upload_results(results, url = "http://www.mas.ncl.ac.uk/~ncsg3/form.php",
  args = NULL)
}
\arguments{
\item{results}{Benchmark results. Probably obtained from 
\code{benchmark_std()}.}

\item{url}{The location of where to upload the results.}

\item{args}{Default \code{NULL}. A list of arguments to 
be passed to \code{get_sys_details}.}
}
\description{
This function uploads the benchmarking results. 
These results will then be incorparated
in future versions of the package.
}

